requirements.txt
pyautogui
pytesseract
Pillow
smtplib
psutil
